Fixed resolutions, moved host to internal, renamed device into fs
