+++ cmd/motor-lib/config/device.go
