+++ docs/blockchain/modules/CHANNEL.md
