+++ docs/blockchain/modules/REGISTRY.md
