+++ internal/host/discover/config.go
