+++ internal/host/exchange/util.go
