+++ internal/host/host.go
@@ -18,8 +18,8 @@ import (
 	dht "github.com/libp2p/go-libp2p-kad-dht"
 	ps "github.com/libp2p/go-libp2p-pubsub"
 	"github.com/libp2p/go-msgio"
+	device "github.com/sonr-io/sonr/pkg/fs"
 	"github.com/sonr-io/sonr/pkg/highway/config"
-	"github.com/sonr-io/sonr/pkg/motor/device"
 	t "go.buf.build/grpc/go/sonr-io/core/types/v1"
 	types "go.buf.build/grpc/go/sonr-io/core/types/v1"
 	"google.golang.org/protobuf/proto"
