+++ internal/host/matrix/protocol.go
@@ -18,7 +18,7 @@ import (
 	us "github.com/matrix-org/dendrite/userapi/storage"
 	"github.com/matrix-org/gomatrixserverlib"
 	"github.com/sirupsen/logrus"
-	"github.com/sonr-io/sonr/pkg/host"
+	"github.com/sonr-io/sonr/internal/host"
 )
 
 const (
