+++ internal/host/matrix/rooms.go
