+++ internal/host/state.go
