--- pkg/motor/bind.go
@@ -1,1 +1,0 @@
-package motor
