+++ web/dashboard/Taskfile.yaml
@@ -3,12 +3,10 @@ version: '3'
 tasks:
   build:
     cmds:
-      - npm install
       - npm run build
 
   start:
     cmds:
-      - npm install
       - npm run start
 
   eject:
@@ -17,5 +15,4 @@ tasks:
 
   test:
     cmds:
-      - npm install
       - npm run test
