+++ web/wallet/Taskfile.yaml
@@ -3,7 +3,6 @@ version: '3'
 tasks:
   build:
     cmds:
-      - npm install
       - npm run build
 
   lint:
@@ -12,7 +11,6 @@ tasks:
 
   dev:
     cmds:
-      - npm install
       - npm run dev
 
   test:
@@ -21,5 +19,4 @@ tasks:
 
   serve:
     cmds:
-      - npm install
       - npm run serve
